# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "CMDB 后端接口文档，包含服务器管理、机房管理、用户管理、告警配置等接口。",
        "title": "CMDB 智能运维平台 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8088",
    "basePath": "/api",
    "paths": {
        "/datacenters": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "机房"
                ],
                "summary": "机房列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DataCenter"
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "本地账号或 LDAP 账号登录，返回 JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "password": {
                                    "type": "string"
                                },
                                "username": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                },
                                "user": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/servers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "支持关键字搜索、状态筛选、机柜筛选",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "服务器列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键字",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态: online/offline",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "机柜ID，unassigned=未分配",
                        "name": "cabinet_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Server"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "创建服务器",
                "parameters": [
                    {
                        "description": "服务器信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Server"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/servers/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "服务器详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Server"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "更新服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Server"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "服务器"
                ],
                "summary": "删除服务器",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "服务器ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "返回 CPU/内存/磁盘/网络流量 Top 20 在线服务器",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "资源统计",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Cabinet": {
            "type": "object",
            "properties": {
                "col": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data_center_id": {
                    "type": "integer"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "pos_x": {
                    "type": "integer"
                },
                "pos_y": {
                    "type": "integer"
                },
                "row": {
                    "type": "string"
                },
                "servers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Server"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.DataCenter": {
            "type": "object",
            "properties": {
                "cabinets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Cabinet"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Server": {
            "type": "object",
            "properties": {
                "agent_version": {
                    "type": "string"
                },
                "cabinet": {
                    "$ref": "#/definitions/models.Cabinet"
                },
                "cabinet_id": {
                    "type": "integer"
                },
                "cpu_cores": {
                    "type": "integer"
                },
                "cpu_model": {
                    "type": "string"
                },
                "cpu_usage": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "disk_total": {
                    "type": "integer"
                },
                "disk_used": {
                    "type": "integer"
                },
                "hostname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "last_report": {
                    "type": "string"
                },
                "manual": {
                    "type": "boolean"
                },
                "mem_total": {
                    "type": "integer"
                },
                "mem_used": {
                    "type": "integer"
                },
                "model": {
                    "type": "string"
                },
                "net_in": {
                    "description": "入流量 bytes/s",
                    "type": "integer"
                },
                "net_out": {
                    "description": "出流量 bytes/s",
                    "type": "integer"
                },
                "os": {
                    "type": "string"
                },
                "remark": {
                    "type": "string"
                },
                "slot": {
                    "description": "机柜位号，0表示未指定",
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                },
                "warranty_end": {
                    "description": "维保到期日，手动填写",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "格式: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}